relocating(routes/userRouter): locating logic
cleaning route code
locating try catch into handler
locating logic into controller
